+++ weasyprint/layout/column.py
@@ -51,11 +51,10 @@ def columns_layout(context, box, bottom_space, skip_stack, containing_block,
     if width == 'auto' and count != 'auto':
         width = max(0, box.width - (count - 1) * gap) / count
     elif width != 'auto' and count == 'auto':
-        count = max(1, int(floor((box.width + gap) / (width + gap))))
+        count = max(1, floor((box.width + gap) / (width + gap)))
         width = (box.width + gap) / count - gap
     else:  # overconstrained, with width != 'auto' and count != 'auto'
-        count = max(
-            1, min(count, int(floor((box.width + gap) / (width + gap)))))
+        count = max(1, min(count, floor((box.width + gap) / (width + gap))))
         width = (box.width + gap) / count - gap
 
     # Handle column-span property with the following structure:
